+++ test/unit/http/proxy/read-test.js
@@ -52,13 +52,13 @@ let staticStub = {
   'index.html': 'index-b2d4f6.html',
   '@noCallThru': true
 }
-let read = proxyquire('../../../../../src/http/proxy/read', {
+let read = proxyquire('../../../../src/http/proxy/read', {
   './sandbox': sandboxStub,
   'fs': {existsSync: () => false},
   'aws-sdk': S3Stub,
 })
 // Could maybe do this all in a single proxyquire, but having static.json appear in separate call adds extra insurance against any inadvertent static asset manifest requiring and default key fallback
-let readStatic = proxyquire('../../../../../src/http/proxy/read', {
+let readStatic = proxyquire('../../../../src/http/proxy/read', {
   './sandbox': sandboxStub,
   'fs': {
     existsSync: () => true,
