+++ test/unit/queues/subscribe-test.js
@@ -1,7 +1,7 @@
 let test = require('tape')
 let sinon = require('sinon')
-let subscribe = require('../../../../src/queues/subscribe')
-let mockSqsEvent = require('../../../mock/mock-sqs-event.json')
+let subscribe = require('../../../src/queues/subscribe')
+let mockSqsEvent = require('../../mock/mock-sqs-event.json')
 
 test('queues.subscribe calls handler', t=> {
   t.plan(1)
